exasock: fix epoll maybe-ready queue cleanup
When a socket gets removed from epoll we need to remove it from maybe-ready
queue straight away instead of relying on "lazy removal" in a next round of
epoll_pwait_spin(). Otherwise scenarios of removing followed by adding of
the same fd value to epoll() may end up in reporting duplicated events for
the socket.