exasock: add SO_EXA_MCAST_LISTEN private socket option
This socket option allows to enable kernel bypass on a multicast socket. It may
be used as an alternative to standard IP_ADD_MEMBERSHIP socket option whenever
joining the multicast group in kernel on that socket is not needed or when
receiving the multicast group's data from any local ExaNIC interface on that
socket is required.