exasock: TCP data processing in kernel
With this change exasock kernel module is going to write received TCP segments
to socket's buffer if only they have not been processed by the user space side
yet.

Whenever a new data is going to be processed either by the library or the kernel
module, it is getting locked by incrementing of the shared proc_seq value. Kernel
locks new data segment for processing only if user space side is not processing
any new data at the moment and the segment is ready to get copied to the socket's
buffer (is not out of order). Out of order segments get temporarily stored by
the module in its local list of skbs.

The library stores all processed segments (in order and out of order) directly
in socket's receive buffer, just as it did so far. As opposed to the kernel
module, the library may lock new segment for processing even when kernel is
already in the middle of processing of a previous segment (it can further
increment proc_seq already incremented by the module).