exasock: add kernel epoll notify for TCP connected sockets
Previously the kernel epoll notify mechanism only supports TCP listening
sockets. This change adds support for normal TCP connections as well.

The ioctl interface is changed so that only a file descriptor of the TCP
socket is required - a file descriptor is sufficient to locate the kernel
exasock_tcp struct.