exasock: fix epoll_ctl_add -> bind -> close -> close hang
Previously exa_notify_remove_sock would error out when exa_notify_kern_epoll_del returns an error.  This breaks in the corner case where the socket is put into bypass but an "kern_epoll" instance is not allocated, such as when bind() is called to bind the socket to an exanic interface.  This commit makes it so that cleaning up the userspace metadata is not skipped even when exa_notify_kern_epoll_del fails.