4.16 and 4.17 kernel panic fix
In Linux kernel 4.16 and 4.17, the function i2c_bit_add_bus attempts to
set the clock and data signals on the i2c bus. However, the exanic
pointer was not filled in until after the bus was registered, resulting
in NULL pointer dereference.

This commit fills the exanic pointer into the i2c adapter data before
registering the bus to work around the issue.