Remove the step to override QSFP module power configuration at port up/down
It is incorrect to power down the whole module when one lane is powered down.
This commit also makes bringing up a port succeed regardless of whether
communication with the QSFP succeeds, in case the customer uses
buggy/incompatible modules - this provides a chance that the QSFP will still
work.