Add support for bonding in exasock.
The bonding support code within the exasock module monitors each bond under
management for link state changes.  This monitor service will attempt to mirror
the frequency of the miimon and arp_monitor services running in the bonding.ko
module for that bond interface.

It exports the current bond state to userspace via a /dev/exabond-* device
which can be mmap()ed into userspace processes.  This per-bond device is used
by libexasock to determine which link within the bond is active, or whether
there is an active link *at all* (it is possible for the entire bond to simply
be down).

/sys nodes are exported to enable the sysadmin to add and remove bonding ifaces
from being under management by the exasock module.