Add a reset_on_remove sysfs file to the ExaNIC driver, and add code to exanic-fwupdate to use it when available.
When reset_on_remove is set to 1, the driver resets the ExaNIC just before
deattaching from the device in remove().  This solves a race condition when
doing this from userspace - userspace requires the driver to be loaded to write
the reset register, but the time taken to subsequently unload the driver is not
well bounded, so the card may reset before the driver unloads causing a fatal
PCIe error (NIC-659).