exasock: fix issue with accelerated connections on Ubuntu 20.xx and upstream kernels 5.4.68+, causing either connection failure or lack of acceleration.
The root cause of this issue: on recent kernels, ip_route_output_key does not
update the floi4_oif field of the flowi4 structure, so the output interface
index was 0 even though we managed to find a route.  The modified code retrieves
the output interface from the dst_entry rather than the flowi4 structure.